Working copy

+++ src/main/java/com/thoughtworks/Battalion.java
@@ -4,7 +4,7 @@ public interface Battalion extends Comparable {
     public String getName();
     public Integer getCount();
     public void setCount(Integer count);
-    public BattalionResponse getBattalionResponse(Integer opposingBattalionCount, Integer relativePower);
+    public BattalionResponse getBattalionResponse(Integer opposingBattalionCount, Double relativePower);
     public Integer getRank();
     public Battalion getBattalion(Integer proposedCount);
 }
